+++ src/main/kotlin/library/NodePackage/BSTNode.kt
@@ -6,4 +6,3 @@ class BSTNode<K : Comparable<K>, T>(
     left: BSTNode<K, T>? = null,
     right: BSTNode<K, T>? = null
 ) : Node<K, T, BSTNode<K, T>>(key, value, left, right)
-
