+++ src/main/kotlin/library/Tree/AVLTree.kt
@@ -1,11 +1,10 @@
 package org.example.library.Tree
 
+import library.Tree.BinaryTree
 import org.example.library.NodePackage.AVLNode
 import kotlin.math.max
 
-public class AVLTree<K : Comparable<K>, T> : BinaryTree<K, T, AVLNode<K, T>> {
-    private var root: AVLNode<K, T>? = null
-    private var size: Int = 0
+public class AVLTree<K : Comparable<K>, T> : BinaryTree<K, T, AVLNode<K, T>>() {
 
     // Получает высоту узла
     private fun getHeight(node: AVLNode<K, T>?): Int {
@@ -118,96 +117,8 @@ public class AVLTree<K : Comparable<K>, T> : BinaryTree<K, T, AVLNode<K, T>> {
         // Балансирует дерево после удаления узла
         updateHeight(node)
         return balanceRotation(node)
-
     }
-
     override fun delete(key: K) {
         root = remove(key, root)
-    }
-
-    override fun find(key: K): T? {
-        return findNode(key)?.value
-    }
-
-    private fun findNode(key: K): AVLNode<K, T>? {
-        if (root == null) {
-            return null
-        } else {
-            var node = root
-            while (node != null) {
-                if (node.key == key) return node
-                else if (node.key > key) {
-                    node = node.left
-                } else {
-                    node = node.right
-                }
-            }
-            return null
-        }
-    }
-
-    private fun minNode(root: AVLNode<K, T>?): AVLNode<K, T>? {
-        var root_copy = root
-        while (root_copy?.left != null)
-            root_copy = root_copy.left
-        return root_copy
-    }
-
-    override fun min(): T? {
-        return minNode(root)?.value
-    }
-
-    private fun maxNode(root: AVLNode<K, T>?): AVLNode<K, T>? {
-        var root_copy = root
-        while (root_copy?.right != null)
-            root_copy = root_copy.right
-        return root_copy
-    }
-    override fun max(): T? {
-        return maxNode(root)?.value
-    }
-
-    override fun printTree() {
-        printTreePrivate(root)
-    }
-
-    private fun printTreePrivate(node: AVLNode<K, T>?){
-        if (node == null) return
-        printTreePrivate(node.left)
-        println("${node.key} : ${node.value}")
-        printTreePrivate(node.right)
-    }
-
-    override fun getSize(): Int {
-        return size
-    }
-
-    inner class Iterate : Iterator<Pair<K, T>> {
-        var array: ArrayDeque<Pair<K, T>> = ArrayDeque()
-        var f: Boolean = true
-
-        override fun next(): Pair<K, T> {
-            return array.removeFirst()
-        }
-
-        override fun hasNext(): Boolean {
-            if (f) {
-                getNode(root)
-                f = false
-            }
-            return array.isNotEmpty()
-        }
-
-        fun getNode(node: AVLNode<K, T>?) {
-            if (node != null) {
-                getNode(node.left)
-                array.add(Pair(node.key, node.value))
-                getNode(node.right)
-            }
-        }
-    }
-
-    operator fun iterator(): Iterate {
-        return this.Iterate()
     }
 }
