+++ src/main/kotlin/library/Tree/BSTree.kt
@@ -1,12 +1,9 @@
 package org.example.library.Tree
 
+import library.Tree.BinaryTree
 import org.example.library.NodePackage.BSTNode
 
-class BSTree<K : Comparable<K>, T> : BinaryTree<K, T, BSTNode<K, T>> {
-    private var root: BSTNode<K, T>? = null
-    private var size: Int = 0
-
-
+class BSTree<K : Comparable<K>, T> : BinaryTree<K, T, BSTNode<K, T>>() {
     private fun add(key: K, value: T, node: BSTNode<K, T>?) {
         if (node != null) {
             if (key < node.key) {
@@ -62,94 +59,5 @@ class BSTree<K : Comparable<K>, T> : BinaryTree<K, T, BSTNode<K, T>> {
 
     override fun delete(key: K): Unit {
         root = remove(key, root)
-    }
-
-    override fun find(key: K): T? {
-        return findNode(key)?.value
-    }
-
-    // поиск узла
-    private fun findNode(key: K): BSTNode<K, T>? {
-        if (root == null) {
-            return null
-        } else {
-            var node = root
-            while (node != null) {
-                if (node.key == key) return node
-                else if (node.key > key) {
-                    node = node.left
-                } else {
-                    node = node.right
-                }
-            }
-            return null
-        }
-    }
-
-    private fun minNode(root: BSTNode<K, T>?): BSTNode<K, T>? {
-        var root_copy = root
-        while (root_copy?.left != null)
-            root_copy = root_copy.left
-        return root_copy
-    }
-
-    override fun min(): T? {
-        return minNode(root)?.value
-    }
-
-    // поиск узла с максимальным узлом
-    private fun maxNode(root: BSTNode<K, T>?): BSTNode<K, T>? {
-        var root_copy = root
-        while (root_copy?.right != null)
-            root_copy = root_copy.right
-        return root_copy
-    }
-
-    override fun max(): T? {
-        return maxNode(root)?.value
-    }
-
-    override fun printTree() {
-        printTreePrivate(root)
-    }
-
-    private fun printTreePrivate(node: BSTNode<K, T>?) {
-        if (node == null) return
-        printTreePrivate(node.left)
-        println("${node.key} : ${node.value}")
-        printTreePrivate(node.right)
-    }
-
-    override fun getSize(): Int {
-        return size
-    }
-
-    inner class Iterate : Iterator<Pair<K, T>> {
-        var array: ArrayDeque<Pair<K, T>> = ArrayDeque()
-        var f: Boolean = true
-
-        override fun next(): Pair<K, T> {
-            return array.removeFirst()
-        }
-
-        override fun hasNext(): Boolean {
-            if (f) {
-                getNode(root)
-                f = false
-            }
-            return array.isNotEmpty()
-        }
-
-        fun getNode(node: BSTNode<K, T>?) {
-            if (node != null) {
-                getNode(node.left)
-                array.add(Pair(node.key, node.value))
-                getNode(node.right)
-            }
-        }
-    }
-
-    operator fun iterator(): Iterate {
-        return this.Iterate()
     }
 }
